move ArrayEditor to its own file
